+++ main/src/main/java/net/oneandone/stool/templates/CookieMode.java
@@ -1,0 +1,5 @@
+package net.oneandone.stool.templates;
+
+public enum CookieMode {
+    OFF, STRICT, LEGACY
+}